Extract moderator access into src/core/moderator-access.ts
Move viewer identity, moderator permission lookups/caches, access
assertions, and hub moderator UI state into a dedicated module.
Typecheck, 167 tests, and build all pass.
